+++ relay/android/src/main/java/io/github/magisk317/relay/android/common/utils/XLog.kt
@@ -1,6 +1,5 @@
 package io.github.magisk317.relay.android.common.utils
 
-import android.util.Log
 import io.github.magisk317.smscode.runtime.contract.logging.LogRoute
 
 object XLog {
@@ -32,48 +31,27 @@ object XLog {
 
     @JvmStatic
     fun v(route: LogRoute, message: String, vararg args: Any?) {
-        RelayLogger.log(Log.VERBOSE, route.id, force = false, sensitive = true, message, *args)
+        RelayLogger.v(route, message, *args)
     }
 
     @JvmStatic
     fun d(route: LogRoute, message: String, vararg args: Any?) {
-        RelayLogger.log(Log.DEBUG, route.id, force = false, sensitive = true, message, *args)
+        RelayLogger.d(route, message, *args)
     }
 
     @JvmStatic
     fun i(route: LogRoute, message: String, vararg args: Any?) {
-        RelayLogger.log(
-            Log.INFO,
-            route.id,
-            force = RelayLogger.defaultForceFor(Log.INFO),
-            sensitive = true,
-            message,
-            *args,
-        )
+        RelayLogger.i(route, message, *args)
     }
 
     @JvmStatic
     fun w(route: LogRoute, message: String, vararg args: Any?) {
-        RelayLogger.log(
-            Log.WARN,
-            route.id,
-            force = RelayLogger.defaultForceFor(Log.WARN),
-            sensitive = true,
-            message,
-            *args,
-        )
+        RelayLogger.w(route, message, *args)
     }
 
     @JvmStatic
     fun e(route: LogRoute, message: String, vararg args: Any?) {
-        RelayLogger.log(
-            Log.ERROR,
-            route.id,
-            force = RelayLogger.defaultForceFor(Log.ERROR),
-            sensitive = true,
-            message,
-            *args,
-        )
+        RelayLogger.e(route, message, *args)
     }
 
     @JvmStatic